+++ src/views/table/TableView.ts
@@ -192,7 +192,7 @@ export class TableView implements SubView {
       plugin: this.plugin,
       statuses: config.statuses,
       priorities: config.priorities,
-      subtaskConnectors: config.tableSubtaskConnectors,
+      showSubtreeConnections: config.showSubtreeConnections,
       state: this.state,
       onRefresh: this.onRefresh,
       onSelectionChange: () => {